changed int crc to unsigned fro ghost too

+++ src/game/client/components/ghost.cpp
@@ -451,9 +451,9 @@ void CGhost::Load(const char* pFilename, int ID)
 		unsigned char aSize[4];
 		if(io_read(File, aSize, sizeof(aSize)) != sizeof(aSize))
 			break;
-		int Size = (aSize[0]<<24) | (aSize[1]<<16) | (aSize[2]<<8) | aSize[3];
+		unsigned Size = (aSize[0]<<24) | (aSize[1]<<16) | (aSize[2]<<8) | aSize[3];
 		
-		if(io_read(File, aCompresseddata, Size) != (unsigned)Size)
+		if(io_read(File, aCompresseddata, Size) != Size)
 		{
 			Console()->Print(IConsole::OUTPUT_LEVEL_STANDARD, "ghost", "error reading chunk");
 			break;
